oqs::kem

Struct Kem

pub struct Kem { /* private fields */ }

KEM algorithm

Example

use oqs;
oqs::init();
let kem = oqs::kem::Kem::new(oqs::kem::Algorithm::MlKem512).unwrap();
let (pk, sk) = kem.keypair().unwrap();
let (ct, ss) = kem.encapsulate(&pk).unwrap();
let ss2 = kem.decapsulate(&sk, &ct).unwrap();
assert_eq!(ss, ss2);

Implementations

impl Kem

pub fn new(algorithm: Algorithm) -> Result<Self>

Construct a new algorithm

pub fn algorithm(&self) -> Algorithm

Get the algorithm used by this Kem

pub fn version(&self) -> &'static str

Get the version of the implementation

pub fn claimed_nist_level(&self) -> u8

Get the claimed nist level

pub fn is_ind_cca(&self) -> bool

Is the algorithm ind_cca secure

pub fn length_public_key(&self) -> usize

Get the length of the public key

pub fn length_secret_key(&self) -> usize

Get the length of the secret key

pub fn length_ciphertext(&self) -> usize

Get the length of the ciphertext

pub fn length_shared_secret(&self) -> usize

Get the length of a shared secret

pub fn length_keypair_seed(&self) -> usize

Get the length of a keypair seed

pub fn secret_key_from_bytes<'a>( &self, buf: &'a [u8], ) -> Option<SecretKeyRef<'a>>

Obtain a secret key objects from bytes

Returns None if the secret key is not the correct length.

pub fn public_key_from_bytes<'a>( &self, buf: &'a [u8], ) -> Option<PublicKeyRef<'a>>

Obtain a public key from bytes

Returns None if the public key is not the correct length.

pub fn ciphertext_from_bytes<'a>( &self, buf: &'a [u8], ) -> Option<CiphertextRef<'a>>

Obtain a ciphertext from bytes

Returns None if the ciphertext is not the correct length.

pub fn shared_secret_from_bytes<'a>( &self, buf: &'a [u8], ) -> Option<SharedSecretRef<'a>>

Obtain a secret key from bytes

Returns None if the shared secret is not the correct length.

pub fn keypair_seed_from_bytes<'a>( &self, buf: &'a [u8], ) -> Option<KeypairSeedRef<'a>>

Obtain a keypair seed from bytes

Returns None if the shared secret is not the correct length.

pub fn keypair(&self) -> Result<(PublicKey, SecretKey)>

Generate a new keypair

pub fn keypair_derand<'a, S: Into<KeypairSeedRef<'a>>>( &self, seed: S, ) -> Result<(PublicKey, SecretKey)>

Generate a new keypair from a seed

pub fn encapsulate<'a, P: Into<PublicKeyRef<'a>>>( &self, pk: P, ) -> Result<(Ciphertext, SharedSecret)>

Encapsulate to the provided public key

pub fn decapsulate<'a, 'b, S: Into<SecretKeyRef<'a>>, C: Into<CiphertextRef<'b>>>( &self, sk: S, ct: C, ) -> Result<SharedSecret>

Decapsulate the provided ciphertext