Enum oqs::kem::Algorithm

pub enum Algorithm {

    BikeL1,
    BikeL3,
    ClassicMcEliece348864,
    ClassicMcEliece348864f,
    ClassicMcEliece460896,
    ClassicMcEliece460896f,
    ClassicMcEliece6688128,
    ClassicMcEliece6688128f,
    ClassicMcEliece6960119,
    ClassicMcEliece6960119f,
    ClassicMcEliece8192128,
    ClassicMcEliece8192128f,
    Hqc128,
    Hqc192,
    Hqc256,
    Kyber512,
    Kyber768,
    Kyber1024,
    Kyber512_90s,
    Kyber768_90s,
    Kyber1024_90s,
    NtruHps2048509,
    NtruHps2048677,
    NtruHps4096821,
    NtruHrss701,
    NtruPrimeNtrulpr653,
    NtruPrimeNtrulpr761,
    NtruPrimeNtrulpr857,
    NtruPrimeSntrup653,
    NtruPrimeSntrup761,
    NtruPrimeSntrup857,
    Lightsaber,
    Saber,
    Firesaber,
    FrodoKem640Aes,
    FrodoKem640Shake,
    FrodoKem976Aes,
    FrodoKem976Shake,
    FrodoKem1344Aes,
    FrodoKem1344Shake,
    SidhP434,
    SidhP503,
    SidhP610,
    SidhP751,
    SidhP434Compressed,
    SidhP503Compressed,
    SidhP610Compressed,
    SidhP751Compressed,
    SikeP434,
    SikeP503,
    SikeP610,
    SikeP751,
    SikeP434Compressed,
    SikeP503Compressed,
    SikeP610Compressed,
    SikeP751Compressed,

}

Supported algorithms by OQS

Note that this doesn’t mean that they’ll be available.

Optional support for serde if that feature is enabled.

Variants

BikeL1

BikeL3

ClassicMcEliece348864

ClassicMcEliece348864f

ClassicMcEliece460896

ClassicMcEliece460896f

ClassicMcEliece6688128

ClassicMcEliece6688128f

ClassicMcEliece6960119

ClassicMcEliece6960119f

ClassicMcEliece8192128

ClassicMcEliece8192128f

Hqc128

Hqc192

Hqc256

Kyber512

Kyber768

Kyber1024

Kyber512_90s

Kyber768_90s

Kyber1024_90s

NtruHps2048509

NtruHps2048677

NtruHps4096821

NtruHrss701

NtruPrimeNtrulpr653

NtruPrimeNtrulpr761

NtruPrimeNtrulpr857

NtruPrimeSntrup653

NtruPrimeSntrup761

NtruPrimeSntrup857

Lightsaber

Saber

Firesaber

FrodoKem640Aes

FrodoKem640Shake

FrodoKem976Aes

FrodoKem976Shake

FrodoKem1344Aes

FrodoKem1344Shake

SidhP434

SidhP503

SidhP610

SidhP751

SidhP434Compressed

SidhP503Compressed

SidhP610Compressed

SidhP751Compressed

SikeP434

SikeP503

SikeP610

SikeP751

SikeP434Compressed

SikeP503Compressed

SikeP610Compressed

SikeP751Compressed

Implementations

impl Algorithm

pub fn is_enabled(self) -> bool

Returns true if this algorithm is enabled in the linked version of liboqs

pub fn to_id(self) -> *const c_char

Provides a pointer to the id of the algorithm

For use with the FFI api methods