Enum oqs::sig::Algorithm

pub enum Algorithm {

    Dilithium2,
    Dilithium3,
    Dilithium5,
    Falcon512,
    Falcon1024,
    Picnic3L1,
    Picnic3L3,
    Picnic3L5,
    PicnicL1Fs,
    PicnicL1Ur,
    PicnicL1Full,
    PicnicL3Fs,
    PicnicL3Ur,
    PicnicL3Full,
    PicnicL5Fs,
    PicnicL5Ur,
    PicnicL5Full,
    RainbowIClassic,
    RainbowICircumzenithal,
    RainbowICompressed,
    RainbowIiiClassic,
    RainbowIiiCircumzenithal,
    RainbowIiiCompressed,
    RainbowVClassic,
    RainbowVCircumzenithal,
    RainbowVCompressed,
    SphincsHaraka128fRobust,
    SphincsHaraka128fSimple,
    SphincsHaraka128sRobust,
    SphincsHaraka128sSimple,
    SphincsHaraka192fRobust,
    SphincsHaraka192fSimple,
    SphincsHaraka192sRobust,
    SphincsHaraka192sSimple,
    SphincsHaraka256fRobust,
    SphincsHaraka256fSimple,
    SphincsHaraka256sRobust,
    SphincsHaraka256sSimple,
    SphincsSha256128fRobust,
    SphincsSha256128fSimple,
    SphincsSha256128sRobust,
    SphincsSha256128sSimple,
    SphincsSha256192fRobust,
    SphincsSha256192fSimple,
    SphincsSha256192sRobust,
    SphincsSha256192sSimple,
    SphincsSha256256fRobust,
    SphincsSha256256fSimple,
    SphincsSha256256sRobust,
    SphincsSha256256sSimple,
    SphincsShake256128fRobust,
    SphincsShake256128fSimple,
    SphincsShake256128sRobust,
    SphincsShake256128sSimple,
    SphincsShake256192fRobust,
    SphincsShake256192fSimple,
    SphincsShake256192sRobust,
    SphincsShake256192sSimple,
    SphincsShake256256fRobust,
    SphincsShake256256fSimple,
    SphincsShake256256sRobust,
    SphincsShake256256sSimple,

}

Supported algorithms by liboqs

They may not all be enabled

Optional support for serde if that feature is enabled.

Variants

Dilithium2

Dilithium3

Dilithium5

Falcon512

Falcon1024

Picnic3L1

Picnic3L3

Picnic3L5

PicnicL1Fs

PicnicL1Ur

PicnicL1Full

PicnicL3Fs

PicnicL3Ur

PicnicL3Full

PicnicL5Fs

PicnicL5Ur

PicnicL5Full

RainbowIClassic

RainbowICircumzenithal

RainbowICompressed

RainbowIiiClassic

RainbowIiiCircumzenithal

RainbowIiiCompressed

RainbowVClassic

RainbowVCircumzenithal

RainbowVCompressed

SphincsHaraka128fRobust

SphincsHaraka128fSimple

SphincsHaraka128sRobust

SphincsHaraka128sSimple

SphincsHaraka192fRobust

SphincsHaraka192fSimple

SphincsHaraka192sRobust

SphincsHaraka192sSimple

SphincsHaraka256fRobust

SphincsHaraka256fSimple

SphincsHaraka256sRobust

SphincsHaraka256sSimple

SphincsSha256128fRobust

SphincsSha256128fSimple

SphincsSha256128sRobust

SphincsSha256128sSimple

SphincsSha256192fRobust

SphincsSha256192fSimple

SphincsSha256192sRobust

SphincsSha256192sSimple

SphincsSha256256fRobust

SphincsSha256256fSimple

SphincsSha256256sRobust

SphincsSha256256sSimple

SphincsShake256128fRobust

SphincsShake256128fSimple

SphincsShake256128sRobust

SphincsShake256128sSimple

SphincsShake256192fRobust

SphincsShake256192fSimple

SphincsShake256192sRobust

SphincsShake256192sSimple

SphincsShake256256fRobust

SphincsShake256256fSimple

SphincsShake256256sRobust

SphincsShake256256sSimple

Implementations

impl Algorithm

pub fn is_enabled(self) -> bool

Returns true if this algorithm is enabled in the linked version of liboqs

pub fn to_id(self) -> *const c_char

Provides a pointer to the id of the algorithm

For use with the FFI api methods

pub fn name(&self) -> &'static str

Returns the algorithm’s name as a static Rust string.

This is the same as the to_id, but as a safe Rust string.